// src/types/DocumentEntityType.ts
export enum DocumentEntityType {
  INSTAGRAM_CAPTION = "INSTAGRAM_CAPTION",
  INSTAGRAM_POST_IDEA = "INSTAGRAM_POST_IDEA",
  INSTAGRAM_BIO_DESCRIPTION = "INSTAGRAM_BIO_DESCRIPTION",
  INSTAGRAM_STORY_IDEA = "INSTAGRAM_STORY_IDEA",
  INSTAGRAM_REEL_IDEA = "INSTAGRAM_REEL_IDEA",
  INSTAGRAM_REEL_SCRIPT = "INSTAGRAM_REEL_SCRIPT",

  FACEBOOK_POST_CAPTION = "FACEBOOK_POST_CAPTION",
  FACEBOOK_POST_IDEA = "FACEBOOK_POST_IDEA",
  FACEBOOK_PAGE_ABOUT = "FACEBOOK_PAGE_ABOUT",

  LINKEDIN_POST_CAPTION = "LINKEDIN_POST_CAPTION",
  LINKEDIN_POST_IDEA = "LINKEDIN_POST_IDEA",
  LINKEDIN_PAGE_ABOUT = "LINKEDIN_PAGE_ABOUT",

  TWITTER_TWIT = "TWITTER_TWIT",
  THREADS_TWIT = "THREADS_TWIT",

  EMAIL_CONTENT = "EMAIL_CONTENT",
  EMAIL_IDEA = "EMAIL_IDEA",

  SMS_CONTENT = "SMS_CONTENT",
  SMS_IDEA = "SMS_IDEA",

  PHONE_CALL_SCRIPT = "PHONE_CALL_SCRIPT",

  GOOGLE_ADS_KEYWORDS = "GOOGLE_ADS_KEYWORDS",
  GOOGLE_ADS_IDEA = "GOOGLE_ADS_IDEA",

  ARTICLE = "ARTICLE",
  WEBSITE_ABOUT = "WEBSITE_ABOUT",
  WEBSITE_FAQ = "WEBSITE_FAQ",
  WEBSITE_TESTEMONIAL = "WEBSITE_TESTEMONIAL",
  WEBSITE_PRODUCT_DESCRIPTION = "WEBSITE_PRODUCT_DESCRIPTION",
}